+++ backend/apid/handlers/create_test.go
@@ -3,6 +3,7 @@ package handlers
 import (
 	"bytes"
 	"context"
+	"errors"
 	"net/http"
 	"testing"
 
@@ -53,7 +54,7 @@ func TestHandlers_CreateResource(t *testing.T) {
 			body: marshal(t, fixture.Resource{ObjectMeta: corev2.ObjectMeta{}}),
 			storeFunc: func(s *mockstore.MockStore) {
 				s.On("CreateResource", mock.Anything, mock.AnythingOfType("*fixture.Resource")).
-					Return(&store.ErrNotValid{})
+					Return(&store.ErrNotValid{Err: errors.New("error")})
 			},
 			wantErr: true,
 		},