+++ token/token_test.go
@@ -1,7 +1,8 @@
-package agent
+package token
 
 import (
 	"encoding/json"
+	"reflect"
 	"testing"
 
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
@@ -10,7 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestTokenSubstitution(t *testing.T) {
+func TestSubstitution(t *testing.T) {
 	testCases := []struct {
 		name            string
 		data            interface{}
@@ -120,7 +121,7 @@ func TestTokenSubstitution(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			result, err := TokenSubstitution(dynamic.Synthesize(tc.data), tc.input)
+			result, err := Substitution(dynamic.Synthesize(tc.data), tc.input)
 			testutil.CompareError(err, tc.expectedError, t)
 
 			if !tc.expectedError {
@@ -134,7 +135,7 @@ func TestTokenSubstitution(t *testing.T) {
 	}
 }
 
-func TestTokenSubstitutionLabels(t *testing.T) {
+func TestSubstitutionLabels(t *testing.T) {
 	data := corev2.Check{
 		ObjectMeta: corev2.ObjectMeta{
 			Labels: map[string]string{"foo": "bar"},
@@ -147,7 +148,7 @@ func TestTokenSubstitutionLabels(t *testing.T) {
 			},
 		},
 	}
-	result, err := TokenSubstitution(dynamic.Synthesize(data), input)
+	result, err := Substitution(dynamic.Synthesize(data), input)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -157,5 +158,121 @@ func TestTokenSubstitutionLabels(t *testing.T) {
 	}
 	if got, want := check.Labels["foo"], "bar"; got != want {
 		t.Fatalf("bad sub: got %q, want %q", got, want)
+	}
+}
+
+func TestSubstituteAsset(t *testing.T) {
+	tests := []struct {
+		name    string
+		asset   *corev2.Asset
+		entity  *corev2.Entity
+		wantErr bool
+		want    *corev2.Asset
+	}{
+		{
+			name:  "A token in the URL can be substituted",
+			asset: &corev2.Asset{URL: "{{ .labels.asset_url }}/asset.tar.gz"},
+			entity: &corev2.Entity{ObjectMeta: corev2.ObjectMeta{
+				Labels: map[string]string{"asset_url": "http://127.0.0.1"},
+			}},
+			want: &corev2.Asset{URL: "http://127.0.0.1/asset.tar.gz"},
+		},
+		{
+			name:  "An asset checksum cannot be substituted",
+			asset: &corev2.Asset{Sha512: "{{ .labels.sha }}"},
+			entity: &corev2.Entity{ObjectMeta: corev2.ObjectMeta{
+				Labels: map[string]string{"sha": "83b51af2254470edbeabf840ae556f113452133f4abbe41e0ce5e0ac37d00262a17646d38ddc23fa16f39706f3506ade902eb1b29429bb0898cfd8c5ce0b0e36"},
+			}},
+			want: &corev2.Asset{Sha512: "{{ .labels.sha }}"},
+		},
+		{
+			name:    "Errors encountered while performing token substitution are returned",
+			asset:   &corev2.Asset{URL: "{{ .labels.asset_url }}"},
+			entity:  &corev2.Entity{},
+			wantErr: true,
+			want:    &corev2.Asset{URL: "{{ .labels.asset_url }}"},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if err := SubstituteAsset(tt.asset, tt.entity); (err != nil) != tt.wantErr {
+				t.Errorf("SubstituteAsset() error = %v, wantErr %v", err, tt.wantErr)
+			}
+			if !reflect.DeepEqual(tt.asset, tt.want) {
+				t.Errorf("SubstituteAsset() = %#v, want %#v", tt.asset, tt.want)
+			}
+		})
+	}
+}
+
+func TestSubstituteCheck(t *testing.T) {
+	tests := []struct {
+		name        string
+		check       *corev2.CheckConfig
+		entity      *corev2.Entity
+		wantErr     bool
+		wantCommand string
+	}{
+		{
+			name:  "A token in the command can be substituted",
+			check: &corev2.CheckConfig{Command: "echo {{ .labels.region }}"},
+			entity: &corev2.Entity{ObjectMeta: corev2.ObjectMeta{
+				Labels: map[string]string{"region": "us-west-1"},
+			}},
+			wantCommand: "echo us-west-1",
+		},
+		{
+			name:        "Errors encountered while performing token substitution are returned",
+			check:       &corev2.CheckConfig{Command: "echo {{ .labels.region }}"},
+			entity:      &corev2.Entity{},
+			wantErr:     true,
+			wantCommand: "echo {{ .labels.region }}",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if err := SubstituteCheck(tt.check, tt.entity); (err != nil) != tt.wantErr {
+				t.Errorf("SubstituteCheck() error = %v, wantErr %v", err, tt.wantErr)
+			}
+			if !reflect.DeepEqual(tt.check.Command, tt.wantCommand) {
+				t.Errorf("SubstituteCheck() = %#v, want %#v", tt.check, tt.wantCommand)
+			}
+		})
+	}
+}
+
+func TestSubstituteHook(t *testing.T) {
+	tests := []struct {
+		name        string
+		hook        *corev2.HookConfig
+		entity      *corev2.Entity
+		wantErr     bool
+		wantCommand string
+	}{
+		{
+			name: "A token in the command can be substituted",
+			hook: &corev2.HookConfig{Command: "echo {{ .labels.region }}"},
+			entity: &corev2.Entity{ObjectMeta: corev2.ObjectMeta{
+				Labels: map[string]string{"region": "us-west-1"},
+			}},
+			wantCommand: "echo us-west-1",
+		},
+		{
+			name:        "Errors encountered while performing token substitution are returned",
+			hook:        &corev2.HookConfig{Command: "echo {{ .labels.region }}"},
+			entity:      &corev2.Entity{},
+			wantErr:     true,
+			wantCommand: "echo {{ .labels.region }}",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if err := SubstituteHook(tt.hook, tt.entity); (err != nil) != tt.wantErr {
+				t.Errorf("SubstituteHook() error = %v, wantErr %v", err, tt.wantErr)
+			}
+			if !reflect.DeepEqual(tt.hook.Command, tt.wantCommand) {
+				t.Errorf("SubstituteHook() = %#v, want %#v", tt.hook, tt.wantCommand)
+			}
+		})
 	}
 }